update labs
Lab 1 complete and checked off. Ready for anything within the next two labs. Lab 2 needs a single adjustment to make sure that the duty cycle is correct for both pins. Then, lab3 needs to be completed by going through and adjusting the interrupt service routine for the adc12 items in order to actually store the data.